% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter_question.R
\name{alter_question}
\alias{alter_question}
\title{alter_question function file.}
\usage{
alter_question(prompt, type, options = NULL)
}
\arguments{
\item{prompt}{The text prompt that will be displayed in the app.}

\item{type}{The question type. The function supports the following types: "text", "decimal", "select_one", "likert".}

\item{options}{A vector of possible choices for question types "select_one" and "likert".}
}
\value{
A list containing the question prompt, the question type, and the list of options for "select_one" and "likert" questions.
}
\description{
alter_question function file.
}
