% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_xlsform.R
\name{compile_xlsform}
\alias{compile_xlsform}
\title{compile_xlsform function file.}
\usage{
compile_xlsform(
  layer_list,
  filename = "names.csv",
  type = "jpg",
  photo_confirm = "all"
)
}
\arguments{
\item{layer_list}{A labeled list of network questions to construct.}

\item{filename}{The name of the external csv file with names and PIDS.}

\item{type}{The file extension of the photographs. Should be "jpg" or "png".}

\item{photo_confirm}{A value that specifies whether the photo confirmation group needs to be included in the output. 
Options are "all" (for always), "only_focal" if only confirmation of the focal is needed, and "none" to omit all photo confirmation steps for focal and alters.}
}
\value{
An XLSForm formated "xlxs" file is saved to the working directory This file can be uploaded to KoboCollect.
}
\description{
compile_xlsform function file.
}
