% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_block_to_survey.R
\name{convert_block_to_survey}
\alias{convert_block_to_survey}
\title{Function to deploy a new XLSForm to KoboToolbox.}
\usage{
convert_block_to_survey(url, block_uid, api_token, deploy = TRUE)
}
\arguments{
\item{url}{The base Kobo URL (without https://).}

\item{block_uid}{The uid of the block template.}

\item{api_token}{The API token used to authorize the request.}

\item{deploy}{Deploy the survey.}
}
\value{
Parsed response (list) from Kobo API with form metadata.
}
\description{
Function to deploy a new XLSForm to KoboToolbox.
}
