% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_export.R
\name{create_export}
\alias{create_export}
\title{Function to create a data export file.}
\usage{
create_export(export_url, api_token)
}
\arguments{
\item{export_url}{The url for the export on Kobo.}

\item{api_token}{The token used to sign the request.}
}
\value{
This just pushes file. No return.
}
\description{
Function to create a data export file.
}
