% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_all_files.R
\name{delete_all_files}
\alias{delete_all_files}
\title{Function to delete ALL attached files in a project (caution advised).}
\usage{
delete_all_files(media_url, api_token)
}
\arguments{
\item{media_url}{The url for the media files on Kobo.}

\item{api_token}{The token used to sign the request.}
}
\value{
This just deletes files. No return.
}
\description{
Function to delete ALL attached files in a project (caution advised).
}
