% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files.R
\name{delete_files}
\alias{delete_files}
\title{Function to delete files that are present in the media_files vector.}
\usage{
delete_files(media_url, media_files, api_token)
}
\arguments{
\item{media_url}{The url for the media files on Kobo.}

\item{media_files}{The names of media files on Kobo.}

\item{api_token}{The token used to sign the request.}
}
\value{
This just deletes files. No return.
}
\description{
Function to delete files that are present in the media_files vector.
}
