% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_form.R
\name{deploy_form}
\alias{deploy_form}
\title{Function to deploy or redeploy a KoboToolbox form via API}
\usage{
deploy_form(api_url, asset_id, api_token, redeploy = FALSE)
}
\arguments{
\item{api_url}{The base Kobo API URL (e.g. "https://kf.kobotoolbox.org/api/v2/assets/").}

\item{asset_id}{The UID of the form (asset) to deploy or redeploy.}

\item{api_token}{Your Kobo API token.}

\item{redeploy}{Logical; if TRUE, redeploy instead of first deploy.}
}
\value{
TRUE if deployment successful, FALSE otherwise.
}
\description{
Function to deploy or redeploy a KoboToolbox form via API
}
