% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_export.R
\name{download_export}
\alias{download_export}
\title{Function to download a data file.}
\usage{
download_export(export_url, folder_path, api_token, overwrite = FALSE)
}
\arguments{
\item{export_url}{The url for the export on Kobo.}

\item{folder_path}{The path to save a file to.}

\item{api_token}{The token used to sign the request.}

\item{overwrite}{Allow overwriting?.}
}
\value{
This just pushes file. No return.
}
\description{
Function to download a data file.
}
