% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_q.R
\name{extra_q}
\alias{extra_q}
\title{extra_q function file.}
\usage{
extra_q(q_name, q_prompt, q_type, choice_list = NULL, hint = NA, relevant = NA)
}
\arguments{
\item{q_name}{The name of the question (as a string)}

\item{q_prompt}{The prompt of the question (as a string).}

\item{q_type}{Question type as a string. The options are "text", "decimal", "select_one", and "likert".}

\item{choice_list}{A vector of strings that contains the options for a select_one or a likert question.}

\item{hint}{Text to add to hint field.}

\item{relevant}{Text to add to relevant field.}
}
\value{
A row with an XLSForm question.
}
\description{
extra_q function file.
}
