% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_question.R
\name{extra_question}
\alias{extra_question}
\title{extra_question function file.}
\usage{
extra_question(prompt, type, options = NULL)
}
\arguments{
\item{prompt}{The text prompt that will be displayed in the app.}

\item{type}{The question type. The function supports the following types: "text", "decimal", "select_one", "likert", "note", "integer".}

\item{options}{A vector of possible choices for question types "select_one" and "likert".}
}
\value{
A list containing the question prompt, the question type, and the list of options for "select_one" and "likert" questions.
}
\description{
extra_question function file.
}
