% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal_info.R
\name{focal_info}
\alias{focal_info}
\title{focal_info function file.}
\usage{
focal_info(filename, type)
}
\arguments{
\item{filename}{The name of the external csv file with names and PIDS.}

\item{type}{The file extension of the photographs. Should be "jpg" or "png".}
}
\value{
A data frame containing focal information in the XLSForm style.
}
\description{
focal_info function file.
}
