% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow_q.R
\name{follow_q}
\alias{follow_q}
\title{follow_q function file.}
\usage{
follow_q(layer, q_name, q_prompt, q_type, choice_list = NULL, follow_up_type)
}
\arguments{
\item{layer}{The name of the network layer (as a string).}

\item{q_name}{The name of the question (as a string)}

\item{q_prompt}{The prompt of the question (as a string).}

\item{q_type}{Question type as a string. The options are "text", "decimal", "select_one", and "likert".}

\item{choice_list}{A vector of strings that contains the options for a select_one or a likert question.}

\item{follow_up_type}{A string: "all", "external", or "none", which controls which follow-up questions get asked.}
}
\value{
A matrix corresponding to the input question in XLSForm format.
}
\description{
follow_q function file.
}
