% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kobo_token.R
\name{get_kobo_token}
\alias{get_kobo_token}
\title{Function to get a Kobo token.}
\usage{
get_kobo_token(server_id, username, password)
}
\arguments{
\item{server_id}{The url for the base server on Kobo (no https://).}

\item{username}{Username on Kobo.}

\item{password}{Password on Kobo.}
}
\value{
This just pushes file. No return.
}
\description{
Function to get a Kobo token.
}
