% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internationalize_headers.R
\name{internationalize_headers}
\alias{internationalize_headers}
\title{internationalize_headers function file.}
\usage{
internationalize_headers(
  name_focal = "Select name of focal person",
  confirm_identity = "Confirm the identity of the interviewed person",
  name_alter = "Write the name of the person",
  list_individuals = "List individuals",
  another_person = "another person",
  follow_up = "Did you already provide follow-up details about this person?",
  follow_yes = "Yes",
  follow_no = "No",
  follow_intro =
    "Now we have some follow-up questions about the people not on the roster."
)
}
\arguments{
\item{name_focal}{Text to add to relevant field.}

\item{confirm_identity}{Text to add to relevant field.}

\item{name_alter}{Text to add to relevant field.}

\item{list_individuals}{Text to add to relevant field.}

\item{another_person}{Text to add to hint field.}

\item{follow_up}{Text to add to relevant field.}

\item{follow_yes}{Text to add to relevant field.}

\item{follow_no}{Text to add to relevant field.}

\item{follow_intro}{Text to add to relevant field.}
}
\value{
A header list for internationalization of hardcoded prompts.
}
\description{
This function simply lets users pass non-English texts into the appropriate prompts. Simply translate the defaults and then pass as headers into compile_xlsform().
}
