% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_login.R
\name{kobo_login}
\alias{kobo_login}
\title{Function to login to Kobo and get token.}
\usage{
kobo_login(url, username, password = NULL)
}
\arguments{
\item{url}{The url for the Kobo server.}

\item{username}{The Kobo username.}

\item{password}{Password. If NULL, then it can be type interactively to avoid storing the password in the script.}
}
\value{
A token.
}
\description{
Function to login to Kobo and get token.
}
