% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_to_edgelist.R
\name{layer_to_edgelist}
\alias{layer_to_edgelist}
\title{layer_to_edgelist function file.}
\usage{
layer_to_edgelist(d, layer)
}
\arguments{
\item{d}{A repeat group KoboCollect output.}

\item{layer}{The name of the network layer}
}
\value{
A list containing a network edgelist and a hash list for out of roster individuals for the given network layer.
}
\description{
layer_to_edgelist function file.
}
