% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_kobo_urls.R
\name{make_kobo_urls}
\alias{make_kobo_urls}
\title{Function that lists media files associated with a project.}
\usage{
make_kobo_urls(server_id, asset_id)
}
\arguments{
\item{server_id}{The Kobo server url (without https://).}

\item{asset_id}{The asset.}
}
\value{
A data.frame listing the saved objects.
}
\description{
Function that lists media files associated with a project.
}
