% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_layer.R
\name{net_layer}
\alias{net_layer}
\title{net_layer function file.}
\usage{
net_layer(filename, type, layer, layer_question, headers = NULL)
}
\arguments{
\item{filename}{The name of the external csv file with names and PIDS.}

\item{type}{The file extension of the photographs. Should be "jpg" or "png".}

\item{layer}{The shortname for the question or network layer.}

\item{layer_question}{The full question to be read to respodents.}

\item{headers}{Acessory prompts for internationalization.}
}
\value{
A data frame containing the XLSForm style for one network layer.
}
\description{
net_layer function file.
}
