% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_media.R
\name{retrieve_media}
\alias{retrieve_media}
\title{Function that lists media files associated with a project.}
\usage{
retrieve_media(media_url, api_token)
}
\arguments{
\item{media_url}{The url for the media files on Kobo.}

\item{api_token}{The token used to sign the request.}
}
\value{
A data.frame listing the saved objects.
}
\description{
Function that lists media files associated with a project.
}
