% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_all.R
\name{upload_all}
\alias{upload_all}
\title{Function to delete ALL attached files in a project (caution advised).}
\usage{
upload_all(
  media_url,
  media_folder,
  media_files,
  api_token,
  wait_time = 0.1,
  overwrite = FALSE
)
}
\arguments{
\item{media_url}{The url for the media files on Kobo.}

\item{media_folder}{The local path to the media file folder.}

\item{media_files}{The names of the media files.}

\item{api_token}{The token used to sign the request.}

\item{wait_time}{The waitime between requests.}

\item{overwrite}{If TRUE, then deletes and reuploads existing files.}
}
\value{
This just pushes files. No return.
}
\description{
Function to delete ALL attached files in a project (caution advised).
}
