% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_media.R
\name{upload_media}
\alias{upload_media}
\title{Function to upload a single media file.}
\usage{
upload_media(media_url, file_path, api_token)
}
\arguments{
\item{media_url}{The url for the media files on Kobo.}

\item{file_path}{The local path to the media file.}

\item{api_token}{The token used to sign the request.}
}
\value{
This just pushes file. No return.
}
\description{
Function to upload a single media file.
}
