% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_xlsform.R
\name{upload_xlsform}
\alias{upload_xlsform}
\title{Function to upload a new XLSForm to KoboToolbox.}
\usage{
upload_xlsform(url, xlsform_path, api_token, form_name = NULL)
}
\arguments{
\item{url}{The base Kobo API URL.}

\item{xlsform_path}{The local path to the XLSForm.}

\item{api_token}{The API token used to authorize the request.}

\item{form_name}{The name to assign to the survey in Kobo (optional).}
}
\value{
Parsed response (list) from Kobo API with form metadata.
}
\description{
Function to upload a new XLSForm to KoboToolbox.
}
