% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_xlsform.R
\name{compile_xlsform}
\alias{compile_xlsform}
\title{compile_xlsform function file.}
\usage{
compile_xlsform(filename = "names.csv", type = "jpg", layer_list)
}
\arguments{
\item{filename}{The name of the external csv file with names and PIDS.}

\item{type}{The file extension of the photographs. Should be "jpg" or "png".}

\item{layer_list}{A labeled list of network questions to construct.}
}
\value{
An XLSForm formated "xlxs" file is saved to the working directory This file can be uploaded to KoboCollect.
}
\description{
compile_xlsform function file.
}
