% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_xlsform.R
\name{compile_xlsform}
\alias{compile_xlsform}
\title{compile_xlsform function file.}
\usage{
compile_xlsform(
  layer_list,
  filename_roster = "names.csv",
  filename_xlsform = "network_collect.xlsx",
  type = "jpg",
  photo_confirm = "all",
  follow_up_questions = NULL,
  headers = NULL
)
}
\arguments{
\item{layer_list}{A labeled list of network questions to construct.}

\item{filename_roster}{The name of the external csv file with names and PIDS.}

\item{filename_xlsform}{The name of the XLSForm to be created.}

\item{type}{The file extension of the photographs. Should be "jpg" or "png".}

\item{photo_confirm}{A value that specifies whether the photo confirmation group needs to be included in the output.}

\item{follow_up_questions}{A named list of follow-up questions created using the alter_question() function.}

\item{headers}{Acessory prompts for internationalization. Must be NULL, or a 2-list of prompts.
Options are "all" (for always), "only_focal" if only confirmation of the focal is needed, and "none" to omit all photo confirmation steps for focal and alters.}
}
\value{
An XLSForm formated "xlxs" file is saved to the working directory This file can be uploaded to KoboCollect.
}
\description{
compile_xlsform function file.
}
