% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_to_edgelist.R
\name{kobo_to_edgelist}
\alias{kobo_to_edgelist}
\title{kobo_to_edgelist function file.}
\usage{
kobo_to_edgelist(datafile, save = NULL)
}
\arguments{
\item{datafile}{The name of the XLS data object exported from KoboToolbox.}

\item{save}{If not NULL, then R was save the data with name "save".}
}
\value{
A data frame containing a network edgelist.
}
\description{
kobo_to_edgelist function file.
}
