% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_to_edgelist.R
\name{kobo_to_edgelist}
\alias{kobo_to_edgelist}
\title{kobo_to_edgelist function file.}
\usage{
kobo_to_edgelist(path, questions, save = FALSE)
}
\arguments{
\item{path}{The name of the XLS data object exported from KoboToolbox.}

\item{questions}{The labeled list used in the creation of the XLSForm.}

\item{save}{If TRUE, it saves the output as an XLSX file in the working directory.}
}
\value{
A data frame containing a network edgelist, or a list containing a network edgelist and a hash list for out of roster individuals.
}
\description{
kobo_to_edgelist function file.
}
