% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_details.R
\name{layer_details}
\alias{layer_details}
\title{layer_details function file.}
\usage{
layer_details(
  layer_vec,
  follow_up_questions,
  follow_up_type,
  headers,
  skip_repeat_names = TRUE
)
}
\arguments{
\item{layer_vec}{The name of the network layer for which the follow-up questions are deployed.}

\item{follow_up_questions}{Follow-up questions list.}

\item{follow_up_type}{A string: "all", "external", or "none", which controls which follow-up questions get asked.}

\item{headers}{A list containing info need for skip logic.}

\item{skip_repeat_names}{If true, allow user to skip repeated names in follow up.}
}
\value{
A data frame containing a single layer of follow-up questions in an XLSForm repeat group.
}
\description{
layer_details function file.
}
