% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_details.R
\name{layer_details}
\alias{layer_details}
\title{layer_details function file.}
\usage{
layer_details(layer_vec, q_list)
}
\arguments{
\item{layer_vec}{The name of the network layer for which the follow-up questions are deployed.}

\item{q_list}{Follow-up questions list.}
}
\value{
A data frame containing a single layer of follow-up questions in an XLSForm repeat group.
}
\description{
layer_details function file.
}
